requirements.txt
SQLAlchemy~=2.0.22
alembic~=1.12.0
fastapi~=0.104.0
pydantic~=2.4.2
passlib~=1.7.4
starlette~=0.27.0
uvicorn~=0.23.2
jose~=1.0.0
PyJWT~=2.8.0